requirements.txt
pymupdf
langchain
transformers
pillow
torch
numpy
scikit-learn
langchain-community
python-dotenv
ipykernel
google-generativeai
faiss-cpu
streamlit
pytesseract